Create a README-Pt-Br.md for nut
Many people are using Nut for organizing and usb server in Brazil, and many are teaching each other how to use, but there's people who wants to get the files directly from it's origins, so the community made an PT-BR Readme for you.
They would love to have and PT-BR ReadMe.

I revised all the text. It look's well made!